requirements.txt
groq
edge-tts
pyaudio
pygame
python-dotenv
numpy
colorama
scipy
